fix: add unit tests for launchGodotEditor and consolidate headless tests
